feat: rdepot-pip calls pip (currently equal)
rdepot-pip is now equal to pip. Quick interactive test correctly
demonstrates that:
* rdepot-pip correctly installs in the .venv  (if it is active)
* stdout of pip is nicely returned.
* Interactive commands like `uninstall` that ask for comfirmation work!

dev-related stuff:

* Added pip as dev-dependency inorder to have pip available. Since rye
  (correctly) does not have pip get installed in the .venv. But we do
  need it for the dependency.

Note, this does mean we rely on a user installed pip in virtual
environments. Alternative would be to have `pip` be a hard dependency of
rdepot-pip. Which would mean that we install `pip` as a formal
dependency.
However, we indirectly call `pip` in a subprocess. And so we need to
realy anyway on an installed pip. (which almost always the case)

other:
* short cut script `r` for rdepot-pip.